# api/openapi/sentinel-aegis.yaml
openapi: 3.1.0
info:
  title: SentinelAegis API
  description: |
    Multi-agent BEC fraud detection system. Three independent Gemini 1.5 Pro agents 
    analyze wire transfer requests for social engineering, banking detail manipulation, 
    and behavioral timing anomalies. A consensus engine produces a final HALT/REVIEW/APPROVE decision.
  version: 1.0.0
  contact:
    name: SentinelAegis Team
    url: https://github.com/Mutasem-mk4/sentinelAegis
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://sentinelaegis-471764064985.us-central1.run.app
    description: Production (Cloud Run)
  - url: http://localhost:8080
    description: Local development
paths:
...
  /api/analyze:
    post:
      summary: Analyze a demo transaction
      description: |
        Runs all 3 AI agents concurrently on the specified demo transaction 
        and returns the consensus decision. This is the core analysis endpoint.
      operationId: analyzeTransaction
      tags: [Analysis]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [transaction_id]
              properties:
                transaction_id:
                  type: string
                  example: TXN-004
      responses:
        '200':
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '400':
          description: Invalid request
        '404':
          description: Transaction not found
        '429':
          description: Rate limit exceeded
  /api/analyze/custom:
    post:
      summary: Analyze custom transaction data
      description: |
        Accepts user-provided transaction data for ad-hoc BEC analysis.
        All fields except email_body and amount are optional with sensible defaults.
      operationId: analyzeCustom
      tags: [Analysis]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CustomAnalyzeRequest'
      responses:
        '200':
          description: Analysis complete
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '400':
          description: Invalid request (missing email_body or invalid amount)
        '429':
          description: Rate limit exceeded
...
components:
  schemas:
...
    AgentResult:
      type: object
      properties:
        agent_name:
          type: string
          enum: [email_tone, iban_change, timing]
        risk_level:
          type: string
          enum: [HIGH, MEDIUM, LOW]
        confidence:
          type: number
          minimum: 0
          maximum: 1
          example: 0.92
        flags:
          type: array
          items:
            type: string
        explanation:
          type: string
    ConsensusResult:
      type: object
      properties:
        decision:
          type: string
          enum: [HALT, REVIEW, APPROVE]
        risk_score:
          type: integer
          minimum: 0
          maximum: 100
        explanation:
          type: string
        agent_breakdown:
          type: array
          items:
            $ref: '#/components/schemas/AgentResult'
    AnalysisResponse:
      type: object
      properties:
        transaction_id:
          type: string
        consensus:
          $ref: '#/components/schemas/ConsensusResult'
        latency_ms:
          type: integer
        correlation_id:
          type: string
    CustomAnalyzeRequest:
      type: object
      required: [email_body, amount]
      properties:
        email_body:
          type: string
          description: The email text to analyze for BEC indicators
        vendor_name:
          type: string
          default: Unknown Vendor
        amount:
          type: number
          minimum: 0.01
        currency:
          type: string
          default: USD
        current_iban:
          type: string
        previous_iban:
          type: string
        iban_changed_hours_ago:
          type: integer
          default: 0
        requested_at:
          type: string
          description: "HH:MM format"
        typical_window_start:
          type: string
          default: "09:00"
        typical_window_end:
          type: string
          default: "17:00"
...